+++ src/onegov.form/onegov/form/validators.py
@@ -87,5 +87,6 @@ class ExpectedExtensions(WhitelistedMimeType):
     """
 
     def __init__(self, extensions):
-        mimetypes = set(types_map.get(ext, None) for ext in extensions)
+        mimetypes = set(
+            types_map.get('.' + ext.lstrip('.'), None) for ext in extensions)
         super(ExpectedExtensions, self).__init__(whitelist=mimetypes)